converters/MMMLExpressionToEvent: Add option to deactivate defaults
..and also deactivate the default by default. This is a functionality
 inspired by Lilypond, but in MMML this seems to be much more annoying
 and less useful, because in Lilypond parameters aren't only allocated
 by their position, but also by their type and therefore it's easier to
 handle this. In MMML this seems to be rather annoying because for
 instance when using a specific playing indicator, we always need to
 write:

```
seq
    n 3/4 cs pp fermata.type=fermata
    n 1/4 d  pp fermata.type=None
```

So we always need to explicitly de-activate ('de-default') a parameter.
It's still clear that it may be useful to sometimes set defaults, for
instance for dynamics. But for this - it seems - we would need a
different syntax. Maybe we need a magic string for this, for instance
'!':

```
    n 3/4 cs !pp
```

(We shouldn't add a new 'command' to set defaults, because there
aren't any commands in MMML: there are only expressions, and each
expression is one event. Let's keep it like this so that it stays
simple.)

Anyway this patch doesn't something like this: for now we just make
default-values optional to only activate them when they are really
useful. A long term approach adds this special default magic described
above.