MMMLExpressionToEvent: Add option to skip argument with special value '_'
Before this patch all arguments of a MMML expression needed to be
defined in their given order. This means that if a user wanted to
specify the 5th argument of an expression, it also needed to specify the
four previous arguments even if the user didn't care about them. So for
instance when setting the clef on a rest the user had to write:

    r 1/4 pp arpeggio.direction=None clef.name=bass

So duration, volume and playing indicators needed to be defined,
although they didn't matter. This patch adds the new special argument
value `_`. With this, arguments are ignored, so the previous example
could simply be:

    r 1/4 _ _ clef.name=bass