converters/MMMLExpressionToEvent: Support builtin comments
Before this patch comments could only be expressed with the 'mustache'
template language. Now it's possible to directly write comments in MMML
with the '#' syntax. They are very easy to implement and are very
simple, much simpler than a template language. Therefore they should be
directly part of the MMML specification. Also in this way we no longer
depend on a template language for comments, and at least regarding
comments, we could drop the external template language.